refactor: Consolidate settings definitions into registry
