--- tools/server/webui/src/lib/constants/settings-fields.ts
@@ -1,38 +1,0 @@
-import { SETTINGS_KEYS } from './settings-keys';
-
-/**
- * List of all numeric fields in settings configuration.
- * These fields will be converted from strings to numbers during save.
- */
-export const NUMERIC_FIELDS = [
-	SETTINGS_KEYS.TEMPERATURE,
-	SETTINGS_KEYS.TOP_K,
-	SETTINGS_KEYS.TOP_P,
-	SETTINGS_KEYS.MIN_P,
-	SETTINGS_KEYS.MAX_TOKENS,
-	SETTINGS_KEYS.PASTE_LONG_TEXT_TO_FILE_LEN,
-	SETTINGS_KEYS.DYNATEMP_RANGE,
-	SETTINGS_KEYS.DYNATEMP_EXPONENT,
-	SETTINGS_KEYS.TYP_P,
-	SETTINGS_KEYS.XTC_PROBABILITY,
-	SETTINGS_KEYS.XTC_THRESHOLD,
-	SETTINGS_KEYS.REPEAT_LAST_N,
-	SETTINGS_KEYS.REPEAT_PENALTY,
-	SETTINGS_KEYS.PRESENCE_PENALTY,
-	SETTINGS_KEYS.FREQUENCY_PENALTY,
-	SETTINGS_KEYS.DRY_MULTIPLIER,
-	SETTINGS_KEYS.DRY_BASE,
-	SETTINGS_KEYS.DRY_ALLOWED_LENGTH,
-	SETTINGS_KEYS.DRY_PENALTY_LAST_N,
-	SETTINGS_KEYS.AGENTIC_MAX_TURNS,
-	SETTINGS_KEYS.AGENTIC_MAX_TOOL_PREVIEW_LINES
-] as const;
-
-/**
- * Fields that must be positive integers (>= 1).
- * These will be clamped to minimum 1 and rounded during save.
- */
-export const POSITIVE_INTEGER_FIELDS = [
-	SETTINGS_KEYS.AGENTIC_MAX_TURNS,
-	SETTINGS_KEYS.AGENTIC_MAX_TOOL_PREVIEW_LINES
-] as const;
