+++ tools/server/webui/src/lib/stores/settings-referrer.svelte.ts
@@ -1,4 +1,6 @@
-let _url = $state('#/');
+import { SETTINGS_FALLBACK_EXIT_ROUTE } from '$lib/constants';
+
+let _url = $state<string>(SETTINGS_FALLBACK_EXIT_ROUTE);
 
 export const settingsReferrer = {
 	get url() {
