+++ tools/server/webui/src/lib/types/index.ts
@@ -80,7 +80,11 @@ export type {
 	ParameterValue,
 	ParameterRecord,
 	ParameterInfo,
-	SyncableParameter
+	SyncableParameter,
+	SettingsEntry,
+	SettingsSectionTitle,
+	SettingsSectionEntry,
+	SettingsSection,
 } from './settings';
 
 // Common types
